권한, 인증 체크

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
+import shop.mtcoding.blog._core.errors.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -19,6 +21,13 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable int id, BoardRequest.UpdateDTO reqDTO){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정할 권한이 없습니다.");
+        }
+
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/" + id;
     }
@@ -26,12 +35,24 @@ public class BoardController {
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest request){
         Board board = boardRepository.findById(id);
+
+        if (board == null){
+            throw new Exception404("해당 게시글을 찾을 수 없습니다.");
+        }
+
         request.setAttribute("board", board);
         return "board/update-form";
     }
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable int id){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정할 권한이 없습니다.");
+        }
+
         boardRepository.deleteById(id);
         return "redirect:/";
     }
@@ -59,7 +80,17 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+        boolean isOwner = false;
+        // 로그인을 하고, 게시글의 주인이면 isOwner가 true가 된다.
+        if (sessionUser != null){
+            if (sessionUser.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
+
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
         return "board/detail";
     }
